Pull Request for code mergin on first environment (#10)
* Init

First draft for the creation of the environment with some classes and the basic structure for the environment methods

* Update Env_v1

Changed lots of things regarding methods and the environment itself. New env file for another approach to the problem

* Changes on Env_v2

Env v2 is designed following smth like the Tassel Approach. To be further developed,

---------

Signed-off-by: Adam Ross <14985050+R055A@users.noreply.github.com>
Co-authored-by: Adam Ross <14985050+R055A@users.noreply.github.com>